+++ packages/atomic/cypress/e2e/result-list/result-components/result-link.cypress.ts
@@ -87,7 +87,7 @@ describe('Result Link Component', () => {
       });
 
       it('should render an "atomic-result-text" component containing the title', () => {
-        ResultLinkSelectors.firstInResult().first().should('have.text', title);
+        ResultLinkSelectors.firstInResult().first().find('atomic-text').shadow().should('have.text', title);
       });
     });