+++ packages/atomic/src/components/insight/atomic-insight-facet/atomic-insight-facet.tsx
@@ -28,7 +28,7 @@ import {FacetInfo} from '../../common/facets/facet-common-store';
 import {FacetContainer} from '../../common/facets/facet-container/stencil-facet-container';
 import {FacetGuard} from '../../common/facets/stencil-facet-guard';
 import {FacetHeader} from '../../common/facets/facet-header/stencil-facet-header';
-import {FacetPlaceholder} from '../../common/facets/facet-placeholder/facet-placeholder';
+import {FacetPlaceholder} from '../../common/facets/facet-placeholder/stencil-facet-placeholder';
 import {announceFacetSearchResultsWithAriaLive} from '../../common/facets/facet-search/facet-search-aria-live';
 import {FacetShowMoreLess} from '../../common/facets/facet-show-more-less/stencil-facet-show-more-less';
 import {