+++ packages/atomic/src/components/search/facets/atomic-numeric-facet/atomic-numeric-facet.tsx
@@ -37,7 +37,7 @@ import {FacetContainer} from '../../../common/facets/facet-container/stencil-fac
 import {FacetGuard} from '../../../common/facets/stencil-facet-guard';
 import {FacetHeader} from '../../../common/facets/facet-header/stencil-facet-header';
 import {NumberInputType} from '../../../common/facets/facet-number-input/number-input-type';
-import {FacetPlaceholder} from '../../../common/facets/facet-placeholder/facet-placeholder';
+import {FacetPlaceholder} from '../../../common/facets/facet-placeholder/stencil-facet-placeholder';
 import {formatHumanReadable} from '../../../common/facets/numeric-facet/formatter';
 import {NumericFacetValueLink} from '../../../common/facets/numeric-facet/stencil-value-link';
 import {NumericFacetValuesContainer} from '../../../common/facets/numeric-facet/stencil-values-container';