+++ packages/atomic-react/CHANGELOG.md
@@ -1,3 +1,14 @@
+## <small>3.3.21 (2025-10-30)</small>
+
+* chore(atomic): migrate `atomic-html` (#6202) ([d204f17](https://github.com/coveo/ui-kit/commits/d204f17)), closes [#6202](https://github.com/coveo/ui-kit/issues/6202)
+* chore(atomic): migrate atomic-result-list (#6189) ([eace955](https://github.com/coveo/ui-kit/commits/eace955)), closes [#6189](https://github.com/coveo/ui-kit/issues/6189)
+* [Version Bump][skip ci]: ui-kit publish ([b33d02f](https://github.com/coveo/ui-kit/commits/b33d02f))
+* chore (atomic): migrated atomic-result-children-template to lit (#6199) ([c0f6684](https://github.com/coveo/ui-kit/commits/c0f6684)), closes [#6199](https://github.com/coveo/ui-kit/issues/6199)
+* fix(deps): update react j:kit-282 (#6192) ([183b449](https://github.com/coveo/ui-kit/commits/183b449)), closes [#6192](https://github.com/coveo/ui-kit/issues/6192)
+* ci: migrate to pnpm (#6135) ([00e44cd](https://github.com/coveo/ui-kit/commits/00e44cd)), closes [#6135](https://github.com/coveo/ui-kit/issues/6135)
+
+
+
 ## <small>3.3.21 (2025-10-29)</small>
 
 * fix(deps): update react j:kit-282 (#6192) ([183b449](https://github.com/coveo/ui-kit/commits/183b449)), closes [#6192](https://github.com/coveo/ui-kit/issues/6192)